html/webappapis/dynamic-markup-insertion/opening-the-input-stream/origin-check-in-document-open.html

<!doctype html>
<title>Origin check in document.open()</title>
<link rel="author" title="Jochen Eisinger" href="mailto:jochen@chromium.org">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#opening-the-input-stream">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/resources/common.js"></script>
<body>
<script>
testInIFrame(undefined, (ctx) => {
  try {
    ctx.iframes[0].contentDocument.open();
  } catch (e) {
    assert_unreached("Opening a same origin document throws");
  }
}, "It should be possible to open same origin documents.");

testInIFrame("resources/set-document-domain.html", (ctx) => {
  try {
    document.domain = document.domain;
    ctx.iframes[0].contentDocument.open();
  } catch (e) {
    assert_unreached("Opening a same origin document that has its document.domain set throws");
  }
}, "It should be possible to open same origin documents when setting document.domain.");

testInIFrame("http://{{host}}:{{ports[http][1]}}/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/resources/set-document-domain.svg", (ctx) => {
  document.domain = document.domain;
  // getSVGDocument() only checks same origin-domain.
  const svgdoc = ctx.iframes[0].getSVGDocument();
  // open() however requires same origin.
  assert_throws("InvalidStateError", svgdoc.open.bind(svgdoc),
    "Opening a same origin-domain document didn't throw");
}, "It should not be possible to open an same origin-domain document.");
</script>
</body>
